画像アップロードの部分をコンポーネントで分離　firebaseの設定ファイルを追加
