+++ firebaseInit.js
@@ -1,0 +1,6 @@
+import firebase from 'firebase';
+import 'firebase/firestore'
+import firebaseConfig from './firebaseConfig';
+
+const firebaseApp = firebase.initializeApp(firebaseConfig)
+export default firebaseApp.firestore()
