+++ package.json
@@ -11,6 +11,8 @@
   },
   "dependencies": {
     "core-js": "^3.4.4",
+    "dotenv": "^8.2.0",
+    "firebase-tools": "^7.13.0",
     "vue": "^2.6.10",
     "vue-router": "^3.1.3",
     "vuetify": "^2.1.0",
