+++ src/views/workAdd.vue
@@ -1,30 +1,8 @@
 <template>
   <div>
     <v-container>
+      <uploadImage/>
       <v-layout justify-center wrap>
-        <div class="avator">
-          <v-img
-           class="avator_icon"
-           :src="avatorUrl"/>
-          <input
-            type="file"
-            class="avator_input"
-            id="avator-input"
-            accept="image/png,image/jpeg,image/gif"
-            @change="uploadAvator"/>
-          <v-layout justify-center class="mt-12">
-            <v-btn
-              class="avator_botton"
-              @click="selectAvator"
-              text
-            >
-              <v-img
-                src="../../public/images/camera.svg"
-                width="100px"
-              />
-            </v-btn>
-          </v-layout>
-        </div>
         <v-text-field full-width placeholder="タイトル" :rules="titleRules" required></v-text-field>
         <v-text-field full-width placeholder="テーマ" :rules="themeRules" required></v-text-field>
         <v-text-field full-width placeholder="説明" :rules="discriptionRules" required></v-text-field>
@@ -37,27 +15,12 @@
   </div>
 </template>
 <script>
+import uploadImage from '@/components/uploadImage.vue';
 export default {
-  methods: {
-    selectAvator() {
-      const input = document.querySelector('#avator-input');
-      input.click();
-    },
-    uploadAvator(e) {
-      e.preventDefault();
-      //からだったら何もしない
-      if (e.target.files.length === 0) {
-        return null;
-      }
-      const avatorfile =e.target.files[0];
-      //eslint-disable-next-line
-      console.log(avatorfile);
-      //imageurlにtimestamp付与
-      this.avatorUrl = `${this.avatorUrl}?timestamp=${Date.now()}`
-    },
+  components: {
+    uploadImage
   },
   data: () => ({
-    avatorUrl: '/profile.svg',
     title: '',
     titleRules: [
       v => !!v || 'タイトルを入力してください',
@@ -78,9 +41,3 @@ export default {
   })
 };
 </script>
-<style scoped>
-.avator {
-  width: 300px;
-  height: 300px;
-}
-</style>
